Changed catching exceptions in methods CAUSE i can't catch own exception in Xamarin. Small refactoring. Issue when in AddEditItemPage user in first picker select some category and then select it in second picker

+++ NotABook/NotABook/Models/Item.cs
@@ -7,13 +7,11 @@ using NotABook.Models.Exceptions;
 
 namespace NotABook.Models
 {
-    public class Item : ElementOfTheBook 
+    public class Item : ElementOfTheBook
     {
         #region Fields
 
         private string description;
-        
-        //public Book CurrentBook { get; private set; }
 
         #endregion
 
@@ -28,37 +26,69 @@ namespace NotABook.Models
                 if (IsTestingOff)
                     OnPropertyChanged("Description");
             }
-        }        
-        //
+        }
+        
         public ObservableCollection<Category> Categories
         {
             get
             {
-                if (CurrentBook == null)
-                    throw new BookNullException();                
-                ObservableCollection<Category> categories = new ObservableCollection<Category>();             
-                foreach (CategoryInItem pair in CurrentBook.CategoryInItemsOfBook)
-                {                  
-                    if (pair.GetItemId == Id)
-                    {                 
-                        categories.Add(pair.Category);                     
+                if (IsTestingOff)
+                {
+                    if (CurrentBook == null)
+                        return null;
+
+                    ObservableCollection<Category> categories = new ObservableCollection<Category>();
+                    foreach (CategoryInItem pair in CurrentBook.CategoryInItemsOfBook)
+                    {
+                        if (pair.GetItemId == Id)
+                        {
+                            categories.Add(pair.Category);
+                        }
                     }
+                    return categories;
                 }
-                return categories;
+                else
+                {
+                    if (CurrentBook == null)
+                        throw new BookNullException();
+                    ObservableCollection<Category> categories = new ObservableCollection<Category>();
+                    foreach (CategoryInItem pair in CurrentBook.CategoryInItemsOfBook)
+                    {
+                        if (pair.GetItemId == Id)
+                        {
+                            categories.Add(pair.Category);
+                        }
+                    }
+                    return categories;
+                }
             }
             set
             {
-                if (CurrentBook == null)
-                    throw new BookNullException();
-
-                CategoryInItem.DeleteAllConnectionWithItem(CurrentBook, this);
-                foreach (var category in value ?? throw new ArgumentNullException())
-                {
-                    CategoryInItem.CreateCategoryInItem(CurrentBook, category, this);
-                }
-
                 if (IsTestingOff)
-                    OnPropertyChanged("Categories");
+                {
+                    if (CurrentBook == null || value == null)
+                        return;
+
+                    CategoryInItem.DeleteAllConnectionWithItem(CurrentBook, this);
+                    foreach (var category in value)
+                    {
+                        CategoryInItem.CreateCategoryInItem(CurrentBook, category, this);
+                    }
+
+                    if (IsTestingOff)
+                        OnPropertyChanged("Categories");
+                }
+                else
+                {
+                    if (CurrentBook == null)
+                        throw new BookNullException();
+
+                    CategoryInItem.DeleteAllConnectionWithItem(CurrentBook, this);
+                    foreach (var category in value ?? throw new ArgumentNullException())
+                    {
+                        CategoryInItem.CreateCategoryInItem(CurrentBook, category, this);
+                    }
+                }
             }
         }
 
@@ -66,9 +96,9 @@ namespace NotABook.Models
         #endregion
 
         #region Constr
-        //
+        
         public Item(Book curBook) : base(curBook)
-        {            
+        {
             curBook.ItemsOfBook.Add(this);
         }
 
@@ -85,33 +115,78 @@ namespace NotABook.Models
         public Item(Book curBook, string title, string descriprion, ObservableCollection<Category> categories) : this(curBook, title, descriprion)
         {
             Categories = categories;
-        }       
+        }
         #endregion
 
         #region Methods
-        //
+        
+        public string ChangeBookStr(Book book)
+        {
+            if (CurrentBook == null)
+                return "Current book is null";
+            if (book == null)
+                return "new book is null";
+            return $"Changing {CurrentBook.Title} on {book.Title} is {ChangeBook(book)}";
+        }
+        public static string ChangeBookStr(Book newBook, Item item)
+        {
+            if (item == null)
+                return "Item is null";
+            if (item.CurrentBook == null)
+                return "Current book is null";
+            if (newBook == null)
+                return "new book is null";
+            return $"Changing {item.CurrentBook.Title} on {newBook.Title} is {ChangeBook(newBook, item)}";
+        }
+
         public bool ChangeBook(Book newBook)
         {
-            Book lastBook = this.CurrentBook ?? throw new BookNullException();            
-            lastBook.DeleteItem(this);            
-            CurrentBook = newBook ?? throw new BookNullException();
-            CurrentBook.ItemsOfBook.Add(this);            
+            Book lastBook = null;
+            if (BaseClass.IsTestingOff)
+            {
+                if (CurrentBook == null || newBook == null)
+                    return false;
+                lastBook = this.CurrentBook;
+                lastBook.DeleteItem(this);
+                CurrentBook = newBook;
+            }
+            else
+            {
+                lastBook = this.CurrentBook ?? throw new BookNullException();
+                lastBook.DeleteItem(this);
+                CurrentBook = newBook ?? throw new BookNullException();                
+            }                       
+           
+            CurrentBook.ItemsOfBook.Add(this);
             return !lastBook.ItemsOfBook.Contains(this) && newBook.ItemsOfBook.Contains(this);
         }
         public static bool ChangeBook(Book newBook, Item item)
         {
-            if (item == null)
-                throw new ItemNullException();
-
-            Book lastBook = item.CurrentBook ?? throw new BookNullException();
-            lastBook.DeleteItem(item);
-            item.CurrentBook = newBook ?? throw new BookNullException();
+            Book lastBook = null;
+            if (BaseClass.IsTestingOff)
+            {
+                if (item == null || item.CurrentBook == null || newBook == null)
+                    return false;
+                lastBook = item.CurrentBook;
+                lastBook.DeleteItem(item);
+                item.CurrentBook = newBook;
+            }
+            else
+            {
+                if (item == null)
+                    throw new ItemNullException();
+                lastBook = item.CurrentBook ?? throw new BookNullException();
+                lastBook.DeleteItem(item);
+                item.CurrentBook = newBook ?? throw new BookNullException();
+            }            
+                  
             item.CurrentBook.ItemsOfBook.Add(item);
             return !lastBook.ItemsOfBook.Contains(item) && newBook.ItemsOfBook.Contains(item);
         }
 
         public string GetCategoriesInString()
         {
+            if (Categories == null) return "null";
             if (Categories.Count < 1) return "No one categories";
             StringBuilder stringBuilder = new StringBuilder();
             foreach (Category categories in Categories)
@@ -120,30 +195,62 @@ namespace NotABook.Models
             }
             return stringBuilder.Remove(stringBuilder.Length - 2, 2).ToString();
         }
-        //
-        public bool DeleteItem()
+        
+        public string DeleteItemStr()
         {
             if (CurrentBook == null)
-                throw new BookNullException();
-            
+                return $"Current book of {Title} is null";
+            return $"Deleting of {Title} is {DeleteItem().ToString()}";
+        }
+        public static string DeleteItemStr(Item item)
+        {
+            if (item == null)
+                return "Item is null";
+            if (item.CurrentBook == null)
+                return $"Current book of {item.Title} is null";
+            return $"Deleting of {item.Title} is {DeleteItem(item).ToString()}";
+        }
+
+        public bool DeleteItem()
+        {
+            if (BaseClass.IsTestingOff)
+            {
+                if (CurrentBook == null)
+                    return false;
+            }
+            else
+            {
+                if (CurrentBook == null)
+                    throw new BookNullException();
+            }
+
             CurrentBook.ItemsOfBook.Remove(this);
             CategoryInItem.DeleteAllConnectionWithItem(CurrentBook, this);
 
-            if(IsTestingOff)
+            if (IsTestingOff)
                 CurrentBook.OnPropertyChanged("DateOfLastChanging");
+      
             return !CurrentBook.ItemsOfBook.Contains(this) && !CategoryInItem.IsItemHasConnection(CurrentBook, this);
-        }              
-        public static  bool DeleteItem(Item item)
+        }
+        public static bool DeleteItem(Item item)
         {
-            if (item == null)
-                throw new ItemNullException();
-            if (item.CurrentBook == null)
-                throw new BookNullException();
+            if (BaseClass.IsTestingOff)
+            {
+                if (item == null || item.CurrentBook == null)
+                    return false;
+            }
+            else
+            {
+                if (item == null)
+                    throw new ItemNullException();
+                if (item.CurrentBook == null)
+                    throw new BookNullException();
+            }          
 
             item.CurrentBook.ItemsOfBook.Remove(item);
             CategoryInItem.DeleteAllConnectionWithItem(item.CurrentBook, item);
 
-            if (item.IsTestingOff)
+            if (BaseClass.IsTestingOff)
                 item.CurrentBook.OnPropertyChanged("DateOfLastChanging");
 
             return !item.CurrentBook.ItemsOfBook.Contains(item) && !CategoryInItem.IsItemHasConnection(item.CurrentBook, item);
